Extract JSON-RPC messaging to separate module
Move JSON-RPC communication methods from process.rs to new messaging.rs
module to reduce file size. This refactoring improves code organization by
separating messaging concerns from process lifecycle management.

Changes:
- Create adapter/messaging.rs with send_request, send_notification,
  send_request_optional, receive_response_for_request, and
  process_received_message functions
- Update process.rs to delegate messaging functions via
  with_running_transport wrapper
- Reduce process.rs from 477 to 348 lines (129 line reduction)

All quality gates pass: formatting, linting, tests.
